+++ dev/mix/tasks/gpui.test.codegen.ex
@@ -6,6 +6,9 @@ defmodule Mix.Tasks.Gpui.Test.Codegen do
 
   @impl Mix.Task
   def run(_args) do
+    System.put_env("GPUI_CODEGEN_HOST", "gpui_component")
+    System.put_env("GPUI_SKIP_NATIVE", "1")
+    Mix.Task.run("rustq.gen")
     Mix.Task.run("compile")
     Code.require_file("test/test_helper.exs")